requirements.txt
Django==5.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2 